+++ packages/vscode-extension/src/project/DeviceSessionsManager.ts
@@ -23,7 +23,7 @@ import {
   REMOVE,
 } from "../common/State";
 import { createWebSocketDevtoolsServer, DevtoolsServer } from "./devtools";
-import { MetroProvider, SharedMetroProvider } from "./MetroNew";
+import { MetroProvider, SharedMetroProvider } from "./metro";
 
 const LAST_SELECTED_DEVICE_KEY = "last_selected_device";
 const SWITCH_DEVICE_THROTTLE_MS = 300;
