+++ packages/vscode-extension/src/project/devtools.ts
@@ -208,11 +208,6 @@ export class CDPDevtoolsServer extends DevtoolsServer implements Disposable {
   }
 
   private async createConnection() {
-    if (this.connection) {
-      // NOTE: a single `DebugSession` only supports a single devtools connection at a time,
-      // and there's no need to reinitialize after reloads.
-      return;
-    }
     const debugSession = this.debugSession;
     // NOTE: the binding survives JS reloads, and the Devtools frontend will reconnect automatically,
     // so this should not be needed, but because the debugger on Expo Go + Android can break on reloads,
